Add files via upload
The above preview for UI/UX was designed on parking. The first page is the login page which would require login details from the user, once logged in, the user will be redirected directly to the title page with the parking information. The user also has an option for signing up, following which, the user will be redirected to the signup page. Basic detail need to be entered and submitted following which the user will be directed to the title page